Add things for comments and testing of comments
Multiline comments will be removed after this
commit. It is too difficult to allow both
single-line and multiline comments and to get the
parsing of multiline comments right. More
precisely, it is difficult to render multiline
comments and to get the same comment back upon
parsing.

For example, take a comment like this:

/*//hello
*/

When rendered, this internally may have been a
multiline comment with a single-line comment
inside. However, this parses as one multiline
comment.

Getting this sort of thing right will be too
complicated, and it's not worth the time, so
multiline comments are coming out.