Update dependencies to latest versions
Incorporates the latest version of base so that
the package will build with GHC 7.8.  This also
required removing Template Haskell from the tests;
it was causing problems with GHC 7.8--sometimes it
would not compile at all; other times, it would
compile but the test binary would do nothing.

The tests now use Tasty.